+++ packages/rust-bindings/CHANGELOG.md
@@ -1,3 +1,8 @@
 # Changelog
 
-## Unreleased
+## 0.3.0
+
+### Added
+
+- `createCredentialV1`
+- `createIdRequestV1`